fix(build): resolve bugs and compiler warnings across tree
- cdc_ncm: fix kernel panic on NDP32 mode by passing max_ndp_size instead of pointer to skb_put
- mkimage: use intermediate ints in sscanf to avoid memory corruption on uint8_t fields
- wireless_tools: fix memset in iwevent.c to zero full rtnl_handle struct (sizeof(*rth))
- mt76x2: remove stray uninitialized mac_val bitwise operation in ap_cfg.c
- LPRng: initialize pw pointer with static struct in linelist.c to avoid SIGSEGV
- uClibc-ng: cast offset to uint64_t before 32-bit shift in dl-syscall.h
- depmod: touch dummy modules.builtin.modinfo to silence modern kmod warning
- busybox: clean obsolete symbols CONFIG_LOCK, CONFIG_NETMSG, and duplicate CONFIG_DHCP6C
- dropbear: drop -Wsystem-headers to prevent 300+ false-positive warnings from uClibc headers
- parted: use __attribute__((__malloc__)) to prevent collision with gnulib rpl_malloc
- wsdd2: fix %ld to %zu for size_t / strlen arguments in llmnr.c and wsd.c
- httpd: cast const SSL pointer to (SSL *) in SSL_set_options
- iproute2: guard SIOCGSTAMP and SIOCGSTAMPNS against redefinition

+++ trunk/libc/uClibc-ng-1.0.52/ldso/include/dl-syscall.h
@@ -236,7 +236,7 @@ static __always_inline _syscall6(ssize_t, __syscall_pread, int, fd, void *, buf,
 static __always_inline ssize_t
 _dl_pread(int fd, void *buf, size_t count, off_t offset)
 {
-	return __syscall_pread(fd, buf, count, 0, __LONG_LONG_PAIR((offset >> 32), (offset & 0xffffffff)));
+	return __syscall_pread(fd, buf, count, 0, __LONG_LONG_PAIR(((uint64_t)offset >> 32), (offset & 0xffffffff)));
 }
 #else
 static __always_inline _syscall5(ssize_t, __syscall_pread, int, fd, void *, buf,
